updated based on PR feedback

+++ test/module/royalty/RoyaltyERC1155.t.sol
@@ -22,20 +22,11 @@ import {RoyaltyERC1155} from "src/module/token/royalty/RoyaltyERC1155.sol";
 contract RoyaltyExt is RoyaltyERC1155 {}
 
 contract TransferToken {
-    function transferToken(
-        address payable tokenContract,
-        address from,
-        address to,
-        uint256 tokenId,
-        uint256 value
-    ) public {
-        ERC1155Core(tokenContract).safeTransferFrom(
-            from,
-            to,
-            tokenId,
-            value,
-            ""
-        );
+
+    function transferToken(address payable tokenContract, address from, address to, uint256 tokenId, uint256 value)
+        public
+    {
+        ERC1155Core(tokenContract).safeTransferFrom(from, to, tokenId, value, "");
     }
 
     function batchTransferToken(
@@ -45,17 +36,13 @@ contract TransferToken {
         uint256[] memory tokenIds,
         uint256[] memory values
     ) public {
-        ERC1155Core(tokenContract).safeBatchTransferFrom(
-            from,
-            to,
-            tokenIds,
-            values,
-            ""
-        );
+        ERC1155Core(tokenContract).safeBatchTransferFrom(from, to, tokenIds, values, "");
     }
+
 }
 
 contract RoyaltyERC1155Test is Test {
+
     ERC1155Core public core;
 
     RoyaltyExt public moduleImplementation;
@@ -83,10 +70,11 @@ contract RoyaltyERC1155Test is Test {
 
     bytes32 internal evmVersionHash;
 
-    function signMintRequest(
-        MintableERC1155.MintRequestERC1155 memory _req,
-        uint256 _privateKey
-    ) internal view returns (bytes memory) {
+    function signMintRequest(MintableERC1155.MintRequestERC1155 memory _req, uint256 _privateKey)
+        internal
+        view
+        returns (bytes memory)
+    {
         bytes memory encodedRequest = abi.encode(
             typehashMintRequest,
             _req.tokenId,
@@ -100,9 +88,7 @@ contract RoyaltyERC1155Test is Test {
             _req.uid
         );
         bytes32 structHash = keccak256(encodedRequest);
-        bytes32 typedDataHash = keccak256(
-            abi.encodePacked("\x19\x01", domainSeparator, structHash)
-        );
+        bytes32 typedDataHash = keccak256(abi.encodePacked("\x19\x01", domainSeparator, structHash));
 
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(_privateKey, typedDataHash);
         bytes memory sig = abi.encodePacked(r, s, v);
@@ -127,17 +113,12 @@ contract RoyaltyERC1155Test is Test {
         transferTokenContract = new TransferToken();
 
         // install module
-        bytes memory moduleInitializeData = moduleImplementation
-            .encodeBytesOnInstall(owner, 100, address(0));
-        bytes memory mintableModuleInitializeData = mintableModuleImplementation
-            .encodeBytesOnInstall(owner);
+        bytes memory moduleInitializeData = moduleImplementation.encodeBytesOnInstall(owner, 100, address(0));
+        bytes memory mintableModuleInitializeData = mintableModuleImplementation.encodeBytesOnInstall(owner);
 
         vm.startPrank(owner);
         core.installModule(address(moduleImplementation), moduleInitializeData);
-        core.installModule(
-            address(mintableModuleImplementation),
-            mintableModuleInitializeData
-        );
+        core.installModule(address(mintableModuleImplementation), mintableModuleInitializeData);
         vm.stopPrank();
 
         typehashMintRequest = keccak256(
@@ -145,30 +126,15 @@ contract RoyaltyERC1155Test is Test {
         );
         nameHash = keccak256(bytes("MintableERC1155"));
         versionHash = keccak256(bytes("1"));
-        typehashEip712 = keccak256(
-            "EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"
-        );
-        domainSeparator = keccak256(
-            abi.encode(
-                typehashEip712,
-                nameHash,
-                versionHash,
-                block.chainid,
-                address(core)
-            )
-        );
+        typehashEip712 = keccak256("EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)");
+        domainSeparator = keccak256(abi.encode(typehashEip712, nameHash, versionHash, block.chainid, address(core)));
 
         vm.prank(owner);
         core.grantRoles(owner, Role._MINTER_ROLE);
 
         // set up transfer validator
-        mockTransferValidator = ITransferValidator(
-            0x721C0078c2328597Ca70F5451ffF5A7B38D4E947
-        );
-        vm.etch(
-            address(mockTransferValidator),
-            TRANSFER_VALIDATOR_DEPLOYED_BYTECODE
-        );
+        mockTransferValidator = ITransferValidator(0x721C0078c2328597Ca70F5451ffF5A7B38D4E947);
+        vm.etch(address(mockTransferValidator), TRANSFER_VALIDATOR_DEPLOYED_BYTECODE);
     }
 
     /*///////////////////////////////////////////////////////////////
@@ -177,13 +143,10 @@ contract RoyaltyERC1155Test is Test {
 
     function test_state_setDefaultRoyaltyInfo() public {
         address royaltyRecipient = address(0x123);
-        uint256 royaltyBps = 100;
+        uint96 royaltyBps = 100;
 
         vm.prank(owner);
-        RoyaltyExt(address(core)).setDefaultRoyaltyInfo(
-            royaltyRecipient,
-            royaltyBps
-        );
+        RoyaltyExt(address(core)).setDefaultRoyaltyInfo(royaltyRecipient, royaltyBps);
 
         address receiver;
         uint256 royaltyAmount;
@@ -200,10 +163,7 @@ contract RoyaltyERC1155Test is Test {
         // read state from core
         uint256 salePrice = 1000;
         uint256 tokenId = 1;
-        (receiver, royaltyAmount) = RoyaltyExt(address(core)).royaltyInfo(
-            tokenId,
-            salePrice
-        );
+        (receiver, royaltyAmount) = RoyaltyExt(address(core)).royaltyInfo(tokenId, salePrice);
         assertEq(receiver, royaltyRecipient);
         assertEq(royaltyAmount, (salePrice * royaltyBps) / 10_000);
     }
@@ -215,21 +175,14 @@ contract RoyaltyERC1155Test is Test {
 
     function test_state_setRoyaltyInfoForToken() public {
         address defaultRoyaltyRecipient = address(0x123);
-        uint256 defaultRoyaltyBps = 100;
+        uint96 defaultRoyaltyBps = 100;
 
         address customRoyaltyRecipient = address(0x345);
-        uint256 customRoyaltyBps = 200;
+        uint96 customRoyaltyBps = 200;
 
         vm.startPrank(owner);
-        RoyaltyExt(address(core)).setDefaultRoyaltyInfo(
-            defaultRoyaltyRecipient,
-            defaultRoyaltyBps
-        );
-        RoyaltyExt(address(core)).setRoyaltyInfoForToken(
-            10,
-            customRoyaltyRecipient,
-            customRoyaltyBps
-        );
+        RoyaltyExt(address(core)).setDefaultRoyaltyInfo(defaultRoyaltyRecipient, defaultRoyaltyBps);
+        RoyaltyExt(address(core)).setRoyaltyInfoForToken(10, customRoyaltyRecipient, customRoyaltyBps);
         vm.stopPrank();
 
         address receiver;
@@ -251,29 +204,19 @@ contract RoyaltyERC1155Test is Test {
         uint256 salePrice = 1000;
         uint256 tokenId = 1;
 
-        (receiver, royaltyAmount) = RoyaltyExt(address(core)).royaltyInfo(
-            tokenId,
-            salePrice
-        ); // default royalty
+        (receiver, royaltyAmount) = RoyaltyExt(address(core)).royaltyInfo(tokenId, salePrice); // default royalty
         assertEq(receiver, defaultRoyaltyRecipient);
         assertEq(royaltyAmount, (salePrice * defaultRoyaltyBps) / 10_000);
 
         tokenId = 10;
-        (receiver, royaltyAmount) = RoyaltyExt(address(core)).royaltyInfo(
-            tokenId,
-            salePrice
-        ); // custom royalty
+        (receiver, royaltyAmount) = RoyaltyExt(address(core)).royaltyInfo(tokenId, salePrice); // custom royalty
         assertEq(receiver, customRoyaltyRecipient);
         assertEq(royaltyAmount, (salePrice * customRoyaltyBps) / 10_000);
     }
 
     function test_revert_setRoyaltyInfoForToken() public {
         vm.expectRevert(0x82b42900); // `Unauthorized()`
-        RoyaltyExt(address(core)).setRoyaltyInfoForToken(
-            10,
-            address(0x123),
-            100
-        );
+        RoyaltyExt(address(core)).setRoyaltyInfoForToken(10, address(0x123), 100);
     }
 
     /*///////////////////////////////////////////////////////////////
@@ -281,45 +224,30 @@ contract RoyaltyERC1155Test is Test {
     //////////////////////////////////////////////////////////////*/
 
     function test_state_setTransferValidator() public {
-        assertEq(
-            RoyaltyERC1155(address(core)).getTransferValidator(),
-            address(0)
-        );
+        assertEq(RoyaltyERC1155(address(core)).getTransferValidator(), address(0));
 
         // set transfer validator
         vm.prank(owner);
-        RoyaltyERC1155(address(core)).setTransferValidator(
-            address(mockTransferValidator)
-        );
-        assertEq(
-            RoyaltyERC1155(address(core)).getTransferValidator(),
-            address(mockTransferValidator)
-        );
+        RoyaltyERC1155(address(core)).setTransferValidator(address(mockTransferValidator));
+        assertEq(RoyaltyERC1155(address(core)).getTransferValidator(), address(mockTransferValidator));
 
         // set transfer validator back to zero address
         vm.prank(owner);
         RoyaltyERC1155(address(core)).setTransferValidator(address(0));
-        assertEq(
-            RoyaltyERC1155(address(core)).getTransferValidator(),
-            address(0)
-        );
+        assertEq(RoyaltyERC1155(address(core)).getTransferValidator(), address(0));
     }
 
     function test_revert_setTransferValidator_accessControl() public {
         // attemp to set the transfer validator from an unpermissioned actor
         vm.prank(unpermissionedActor);
         vm.expectRevert(0x82b42900); // `Unauthorized()`
-        RoyaltyERC1155(address(core)).setTransferValidator(
-            address(mockTransferValidator)
-        );
+        RoyaltyERC1155(address(core)).setTransferValidator(address(mockTransferValidator));
     }
 
     function test_revert_setTransferValidator_invalidContract() public {
         // attempt to set the transfer validator to an invalid contract
         vm.prank(owner);
-        vm.expectRevert(
-            RoyaltyERC1155.InvalidTransferValidatorContract.selector
-        );
+        vm.expectRevert(RoyaltyERC1155.InvalidTransferValidatorContract.selector);
         RoyaltyERC1155(address(core)).setTransferValidator(address(11_111));
     }
 
@@ -337,13 +265,7 @@ contract RoyaltyERC1155Test is Test {
         vm.prank(owner);
         core.setApprovalForAll(address(transferTokenContract), true);
 
-        transferTokenContract.transferToken(
-            payable(address(core)),
-            owner,
-            permissionedActor,
-            0,
-            1
-        );
+        transferTokenContract.transferToken(payable(address(core)), owner, permissionedActor, 0, 1);
 
         assertEq(1, core.balanceOf(permissionedActor, 0));
 
@@ -355,13 +277,7 @@ contract RoyaltyERC1155Test is Test {
         values[0] = 1;
         values[1] = 1;
 
-        transferTokenContract.batchTransferToken(
-            payable(address(core)),
-            owner,
-            permissionedActor,
-            tokenIds,
-            values
-        );
+        transferTokenContract.batchTransferToken(payable(address(core)), owner, permissionedActor, tokenIds, values);
 
         assertEq(2, core.balanceOf(permissionedActor, 0));
         assertEq(1, core.balanceOf(permissionedActor, 1));
@@ -381,22 +297,14 @@ contract RoyaltyERC1155Test is Test {
 
         // set transfer validator
         vm.prank(owner);
-        RoyaltyERC1155(address(core)).setTransferValidator(
-            address(mockTransferValidator)
-        );
+        RoyaltyERC1155(address(core)).setTransferValidator(address(mockTransferValidator));
 
         // attempt to transfer token from owner to permissionedActor
         vm.prank(owner);
         core.setApprovalForAll(address(transferTokenContract), true);
 
         vm.expectRevert(0xef28f901);
-        transferTokenContract.transferToken(
-            payable(address(core)),
-            owner,
-            permissionedActor,
-            0,
-            1
-        );
+        transferTokenContract.transferToken(payable(address(core)), owner, permissionedActor, 0, 1);
 
         assertEq(0, core.balanceOf(permissionedActor, 0));
 
@@ -409,13 +317,7 @@ contract RoyaltyERC1155Test is Test {
         values[1] = 1;
 
         vm.expectRevert(0xef28f901);
-        transferTokenContract.batchTransferToken(
-            payable(address(core)),
-            owner,
-            permissionedActor,
-            tokenIds,
-            values
-        );
+        transferTokenContract.batchTransferToken(payable(address(core)), owner, permissionedActor, tokenIds, values);
 
         assertEq(0, core.balanceOf(permissionedActor, 0));
         assertEq(0, core.balanceOf(permissionedActor, 1));
@@ -425,40 +327,30 @@ contract RoyaltyERC1155Test is Test {
                         UTILITY FUNCTIONS
     //////////////////////////////////////////////////////////////*/
 
-    function _mintToken(
-        address to,
-        uint256 quantity,
-        uint256 id,
-        bytes32 uid
-    ) internal {
+    function _mintToken(address to, uint256 quantity, uint256 id, bytes32 uid) internal {
         address saleRecipient = address(0x987);
 
         vm.prank(owner);
         MintableERC1155(address(core)).setSaleConfig(saleRecipient);
 
-        MintableERC1155.MintRequestERC1155 memory mintRequest = MintableERC1155
-            .MintRequestERC1155({
-                tokenId: id,
-                startTimestamp: uint48(block.timestamp),
-                endTimestamp: uint48(block.timestamp + 100),
-                recipient: to,
-                quantity: quantity,
-                currency: NATIVE_TOKEN_ADDRESS,
-                pricePerUnit: 0,
-                uid: uid,
-                metadataURI: "https://example.com/"
-            });
+        MintableERC1155.MintRequestERC1155 memory mintRequest = MintableERC1155.MintRequestERC1155({
+            tokenId: id,
+            startTimestamp: uint48(block.timestamp),
+            endTimestamp: uint48(block.timestamp + 100),
+            recipient: to,
+            quantity: quantity,
+            currency: NATIVE_TOKEN_ADDRESS,
+            pricePerUnit: 0,
+            uid: uid,
+            metadataURI: "https://example.com/"
+        });
         bytes memory sig = signMintRequest(mintRequest, ownerPrivateKey);
 
-        MintableERC1155.MintParamsERC1155 memory params = MintableERC1155
-            .MintParamsERC1155(mintRequest, sig, "");
+        MintableERC1155.MintParamsERC1155 memory params = MintableERC1155.MintParamsERC1155(mintRequest, sig, "");
 
         vm.prank(owner);
         core.mint{value: mintRequest.quantity * mintRequest.pricePerUnit}(
-            mintRequest.recipient,
-            mintRequest.tokenId,
-            mintRequest.quantity,
-            abi.encode(params)
+            mintRequest.recipient, mintRequest.tokenId, mintRequest.quantity, abi.encode(params)
         );
     }
 
@@ -471,14 +363,10 @@ contract RoyaltyERC1155Test is Test {
                 break;
             }
             if (
-                keccak256(abi.encode(file)) ==
-                keccak256(abi.encode('evm_version = "cancun"')) ||
-                keccak256(abi.encode(file)) ==
-                keccak256(abi.encode('evm_version = "london"')) ||
-                keccak256(abi.encode(file)) ==
-                keccak256(abi.encode('evm_version = "paris"')) ||
-                keccak256(abi.encode(file)) ==
-                keccak256(abi.encode('evm_version = "shanghai"'))
+                keccak256(abi.encode(file)) == keccak256(abi.encode('evm_version = "cancun"'))
+                    || keccak256(abi.encode(file)) == keccak256(abi.encode('evm_version = "london"'))
+                    || keccak256(abi.encode(file)) == keccak256(abi.encode('evm_version = "paris"'))
+                    || keccak256(abi.encode(file)) == keccak256(abi.encode('evm_version = "shanghai"'))
             ) {
                 break;
             }
@@ -486,4 +374,5 @@ contract RoyaltyERC1155Test is Test {
 
         evmVersionHash = keccak256(abi.encode(file));
     }
+
 }
